+++ flexmeasures/utils/config_defaults.py
@@ -183,7 +183,7 @@ class Config(object):
     FLEXMEASURES_API_DEFAULT_RATE_LIMIT: str = "500 per minute"
     FLEXMEASURES_API_TRIGGER_RATE_LIMIT: str = "10 per 5 minutes"
     FLEXMEASURES_API_RATE_LIMIT_KEY: str = (
-        "account+asset"  # what to count triggers against: "account+asset", "account" or "user"
+        "account"  # what to count triggers against: "account", "account+asset" or "user"
     )
     FLEXMEASURES_JS_VERSIONS: dict = dict(
         vega="5.22.1",